requirements.txt
flask
Flask-SQLAlchemy
flask-login
mysqlclient
